maybe no need to test these edge cases twice?
Instead we can test them on the parent class.
