+++ lib/streamy/json_event.rb
@@ -1,11 +1,7 @@
 module Streamy
   class JsonEvent < Event
     def payload
-      {
-        type: type,
-        body: body,
-        event_time: event_time
-      }.to_json
+      payload_attributes.to_json
     end
   end
 end
