--- test/support/schemas/overidden_priority.avsc
@@ -1,34 +1,0 @@
-{
-   "type": "record",
-   "name": "overidden_priority",
-   "fields": [
-     {
-       "name": "type",
-       "type": "string"
-     },
-     {
-       "name": "event_time",
-       "type": ["null", "string"],
-       "default": null
-     },
-     {
-       "name": "body",
-       "type": {
-         "type": "record",
-         "name": "body",
-         "fields": [
-           {
-             "name": "smoked",
-             "type": ["null", "string"],
-             "default": null
-           },
-           {
-             "name": "streaky",
-             "type": ["null", "string"],
-             "default": null
-           }
-         ]
-       }
-     }
-   ]
-}
